third_party/blink/web_tests/external/wpt/html/webappapis/scripting/event-loops/fully_active_document.window-expected.txt

This is a testharness.js-based test.
FAIL Tasks for documents that are not fully active are stored, and run when the documents becomes fully-active assert_equals: expected (string) "true" but got (object) null
Harness: the test ran to completion.
